ovn-controller: Monitor chassis_private by chassis name.
Remove the use of sbrec_chassis_is_new() for uncommitted records.  This
is not the way IDL *_is_new() functions are supposed to be used.

Note: With this change if the system-id changes there will be a
transient error in ovn-controller due to ovn-controller trying to insert
a new chassis_private record.  This is due to the fact that the view of
the chassis_private table changes and only chassis_private records
matching the new chassis name are sent to ovn-controller.  This gets
corrected though in the next iteration of the ovn-controller processing
loop.

Suggested-by: Han Zhou <hzhou@ovn.org>
Reported-at: https://mail.openvswitch.org/pipermail/ovs-dev/2020-October/376339.html
Fixes: dce1af31b550 ("chassis: Fix chassis_private record updates when the system-id changes.")
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Acked-by: Mark Gray <mark.d.gray@redhat.com>
Signed-off-by: Han Zhou <hzhou@ovn.org>
(cherry picked from commit 1f915da95dc725131b7df094d494af9fda88ea92)
Acked-by: Vladislav Odintsov <odivlad@gmail.com>
Signed-off-by: Vladislav Odintsov <odivlad@gmail.com>
Signed-off-by: Numan Siddique <numans@ovn.org>