ovn-controller: Remove has_local_l3gateway from struct local_datapath.
Initially the field was added to help update_ct_zones(), but later in
the commit [0] the check was removed.

Now this field is only checked in pinctrl.c, but the condition is in
fact redundant. So this field is not really useful anymore. Remove all
the related code to simplify.

[0] - 41b8c6a0fd ("ovn: Gratuitous ARP for centralized NAT rules on a distributed router")

Acked-by: Numan Siddique <numans@ovn.org>
Signed-off-by: Han Zhou <hzhou@ovn.org>