CI: Use ovn-k8s master commit for k8s kind tests.
All the ovn-k8s tests in OVN are now failing since a few weeks.
Bumping the ovn-k8s commit to the latest one in my private
CI run passed all the tests except the control plane test [1].
control plane test is failing since a while and this needs to be
debugged.

ovn-k8s repo runs the same k8s kind tests with OVN version present in
fedora 33 before merging a patch.  So I think it should be fine to switch
to the master commit of ovn-k8s.

[1] - https://github.com/numansiddique/ovn/runs/3315367278

Signed-off-by: Numan Siddique <numans@ovn.org>
Acked-by: Mark D. Gray <mark.d.gray@redhat.com>