checkpatch: Avoid a case of catastrophic backtracking
The checkpatch script will hang forever on line 282 of a otherwise
valid patch [0].  While the root of the issue is probably
somewhere between the regex itself and the Python re
implementation, this patch provides an early return to avoid this
specific issue.

Also fix the docstring for the if_and_for_end_with_bracket_check
function.

0: https://patchwork.ozlabs.org/project/ovn/patch/20210819110857.2229769-8-frode.nordahl@canonical.com/
Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>
Acked-by: Mark Michelson <mmichels@redhat.com>
Acked-by: Michael Santana <msantana@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>