ovn-northd-ddlog: Simplify LBVIPWithStatus to include up_backends string.
There was only one use for the 'backends' map, which was to be converted
to a string that listed all the backends that were up, so we might as
well do that at its point of origination.  At the same time, everything
else in LBVIPWithStatus was just a copy of the underlying LBVIP, so we
might as well just reference it.

Only slight improvement to performance.

Signed-off-by: Ben Pfaff <blp@ovn.org>
Acked-by: Mark Michelson <mmichels@redhat.com>