Reduce churn and eliminate Cartesian joins.
`Router` and `Switch` relations used to store associated LB UUIDs,
causing a lot of churn when the set of LBs changed.  We remove this
field from `Router` and `Switch` and introduce separate tables
to track the Router-to-LB and Switch-to-LB relations.

In addition we rewrite several rules that computed a Cartesian product
of LBVIP and `Router` or `Switch` tables.

Signed-off-by: Leonid Ryzhyk <lryzhyk@vmware.com>
Signed-off-by: Ben Pfaff <blp@ovn.org>
Acked-by: Mark Michelson <mmichels@redhat.com>