ic: call ovn_db_run() only when all DBs are connected
Prior to this commit ovn_db_run() function in ovn-ic
was called right after daemon's lock was acquired in SB DB.

Inside ovn_db_run() there is a search for availability zone
in NB DB. If AZ was not found, ovn_db_run returned. So, it was
an implicit requirement for connected NB DB.

But in case, where NB DB was connected and ICNB DB was not,
we could run in situation where in ts_run() logical switches
from NB DB were collected and transit switches from ICNB DB
were not due to not synced ICNB DB. This lead to deletion of
transit logical switches from NB DB.

A normal restart of only one ovn-ic daemon in AZ, could damage
the logical topology: it removed LS (transit), its LSPs.
After ovn-ic is connected to ICNB, it creates logical switch
in NB DB back, but its LSPs are lost and need manual re-addition.

This commit adds explicit requirement for all DBs to be
connected:

- NB DB
- SB DB
- ICNB DB
- ICSB DB

Only when this condition is met, we can safely run ovn-ic sync
logic.

Signed-off-by: Vladislav Odintsov <odivlad@gmail.com>
Signed-off-by: Han Zhou <hzhou@ovn.org>