ovn-northd: Avoid verification of NB_Global.options if nothing changed.
northd constructs transaction for the Northbound database options even
if they don't need to be changed.  This also includes verification of
a current state.  IDL will eventually drop the transaction if it will
not contain any other operations, but if there are some other
operations, this useless update to the same value will be included
along with the verification.  This causes transaction failures because
NB_Global.options can be updated by CMS at the same time.

For example, ovn-kubernetes sets 'options:e2e_timestamp' for it's own
purposes, and if this value will be updated while there is an in-flight
transaction from the northd, transaction will fail the verification and
northd will have to re-try it.

To avoid these issues, updating and verifying options only if needed.

Fixes: b07f1bc3d068 ("Add VXLAN support for non-VTEP datapath bindings")
Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Acked-by: Mark D. Gray <mark.d.gray@redhat.com>
Signed-off-by: Mark Michelson <mmichels@redhat.com>