northd: Change the default value of ignore_lsp_down to true.
The current default behavior is that ARP responder flows for VIFs are
added by northd after the port-binding state is UP, which creates more
trouble than benefit in most use cases. To make the default behavior
desirable for majority of the use cases, set the option ignore_lsp_down
to true by default. This would help saving the control plane cost in
large scale environment, reduce the e2e latency for all flows to be
installed for a VIF, and making the VIF readiness checking more convenient
in test cases and likely in CMS as well. User can still set it to false
in circumstances (if any) when this behavior is not desired.

Signed-off-by: Han Zhou <hzhou@ovn.org>
Acked-by: Mark D. Gray <mark.d.gray@redhat.com>
Acked-by: Numan Siddique <numans@ovn.org>