northd: Use ct_(snat/dnat)_in_czone action for distributed routers.
Make of use of these new actions for the distributed routers
for NAT.  These new actions ensure that both sNAT and dNAT
happens in the same zone.  This approach solves a couple of
problems:

 - The datapath flows generated for external traffic which requires
   dNAT (N -> S) or sNAT (S -> N) are completely HWOL'able.

 - Since there is only one zone, it would avoid multiple recirculations
   (improving the performance).

If the packet needs to be both sNATted and dNATted (for hairpin traffic
with source and destination on the same chassis), then sNAT is done
in a separate zone.  To detect this scenario, this patch adds a few
extra logical flows.  For each dnat_and_snat entry prior to this patch
ovn-northd was generating 9 logical flows and with this patch it now
generates 12 logical flows.

Similar approach can be taken for gateway routers.

Reported-at: https://bugzilla.redhat.com/show_bug.cgi?id=1984953
Acked-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>