lflow-cache: Remove conjunction id cache.
Conjunction id cache is useful to keep conjunction ids consistent between
iterations when match or expr cache is not available. However, the
current implementation caches the conjunction id whenever n_conjs > 0.
It assumes that when the lflow is reprocessed, the conjunction ids
starting from the saved conj_id_ofs to the conj_id_ofs + n_conjs - 1 are
always available. This would be true only if the n_conjs doesn't change.
Unfortunately, the n_conjs can change when a lflow is reprocessed, e.g.
when the members of an address-set/port-group increases from 1 to >1.
This would result in conflict conjunction ids being used by different
lflows. This patch adds a test case to cover this scenario, and the test
passes when conjunction id cache is removed.

A follow-up patch will take a different approach to keep conjunction ids
consistent.

Signed-off-by: Han Zhou <hzhou@ovn.org>
Acked-by: Numan Siddique <numans@ovn.org>