lflow: Consistent conjunction id generation.
When a logical flow translation uses conjunction ids, it is better to
keep the conjuntion ids consistent between iterations, so that the
generated OVS flows don't change unnecessarily and avoid reinstalling
unchanged flows to OVS. The problem was partially solved when
lflow-cache was enabled but there was a bug for a corner case so the
lflow-cache based approach was reverted in a previous commit.

This patch takes an approach that generates conjunction ids based on the
logical flow's uuid, which ensures the same ids being used when
processing the same lflow in most cases. There is a very little chance
that different logial flows happen to have the same 32-bit prefix in
their uuids, but the probability is really low. For the very unlikely
situation, the conflicts are handled properly and correctness is
ensured.

A new module lflow-conj-ids is created to manage the id allocations and
maintaining the usage of each ids and their owner lflows.

Unlike the previously lflow-cache based approach, this approach works
equally well when lflow-cache is not in use.

Signed-off-by: Han Zhou <hzhou@ovn.org>
Acked-by: Numan Siddique <numans@ovn.org>