tests: Really fix requested-chassis localport test
The first test case of adding localport without a
requested-chassis option is invalid as this would make the
chassis do claim/release thrashing leading to a unpredictable
test results.

Create the localport in NB DB prior to creating the ports on the
chassis.  The step by step ovn-nbctl CLI creation of the localport
would in reality create a VIF which is then changed to be a
localport, this does not represent how the CMS would create the
localport.

Create a second port representing a VM to ensure the localport
flows are generated.

Use OVS_WAIT_UNTIL instead of AT_CHECK, which should avoid race
conditions where the check is ran before the flow tables are
populated.

Print output so that it is easier to investigate in the event of a
failure.

Fixes: ad77db239d9a ("controller: Make use of Port_Binding:requested_chassis.")
Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>
Acked-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>