northd: Introduce struct northd_data
'struct northd_data' is used to hold the global state data for the
incremental processing node 'en_northd'. This structure will also hold
'struct northd_input' which will hold references to output data from
its input nodes. In particular, this will hold references to database tables
and indexes. In order to achieve this, we refactor in the following way:

* Introduce northd_init() which initializes this data.
* Introduce northd_destroy() which clears this data for a new iteration.
* Remove 'ovn_internal_version' from the context passed to northd
  as it can be determined within northd using ovn_get_internal_version.
* Remove 'use_parallel_build' from the context as it is read from DB as
  suggested by Han.
* Refactor northd.c to use 'struct northd_data' and 'struct northd_input'
  where applicable.

Signed-off-by: Mark Gray <mark.d.gray@redhat.com>
Signed-off-by: Han Zhou <hzhou@ovn.org>